+++ auditlog_ext_nrq/models/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import auditlog_log
+from . import auditlog_log_line
