+++ auditlog_ext_nrq/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This modules does the following:
+  * Allows to user can check the audit log entry based on user timezone only for datetime field.
